services/api/requirements-prod.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
psycopg[binary]>=3.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
